# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediaVault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /metadata/categories:
    get:
      operationId: getMetadataCategories
      tags: [metadata]
      summary: List all metadata categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetadataCategory"
    post:
      operationId: createMetadataCategory
      tags: [metadata]
      summary: Create a metadata category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataCategory"
  /metadata/categories/{id}:
    patch:
      operationId: updateMetadataCategory
      tags: [metadata]
      summary: Update a metadata category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataCategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataCategory"
    delete:
      operationId: deleteMetadataCategory
      tags: [metadata]
      summary: Delete a metadata category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /metadata/values:
    get:
      operationId: getMetadataValues
      tags: [metadata]
      summary: List metadata values (optionally filtered by category)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MetadataValue"
  /metadata/values/{id}:
    patch:
      operationId: updateMetadataValue
      tags: [metadata]
      summary: Rename a metadata value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataValueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataValue"
    delete:
      operationId: deleteMetadataValue
      tags: [metadata]
      summary: Delete a metadata value
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /metadata/values/merge:
    post:
      operationId: mergeMetadataValues
      tags: [metadata]
      summary: Merge multiple metadata values into one
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeValuesInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataValue"

  # ── Playlists ─────────────────────────────────────────────────────────────
...
  /player/history:
    get:
      operationId: getPlaybackHistory
      tags: [player]
      summary: Get playback history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaybackEntry"
  /player/progress:
    post:
      operationId: savePlaybackProgress
      tags: [player]
      summary: Save playback progress for a media file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackProgressInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackEntry"
  /player/progress/{mediaId}:
    get:
      operationId: getPlaybackProgress
      tags: [player]
      summary: Get saved playback progress for a media file
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackEntry"
        "404":
          description: Not found

  # ── Saved Searches ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MediaFile:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [video, audio]
        size:
          type: integer
        duration:
          type: ["number", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        fps:
          type: ["number", "null"]
        bitrate:
          type: ["integer", "null"]
        codec:
          type: ["string", "null"]
        format:
          type: string
        thumbnail:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        favorite:
          type: boolean
        playCount:
          type: integer
        lastPlayed:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, offline]
        dateAdded:
          type: string
        dateModified:
          type: string
        folderId:
          type: ["integer", "null"]
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/MediaMetadataEntry"
      required: [id, path, filename, type, size, format, favorite, playCount, status, dateAdded, dateModified]
...
    MediaMetadataEntry:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        valueId:
          type: integer
        value:
          type: string
      required: [categoryId, categoryName, categorySlug, valueId, value]
...
    MetadataCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        isDefault:
          type: boolean
        valueCount:
          type: integer
        createdAt:
          type: string
      required: [id, name, slug, isDefault, valueCount, createdAt]
    MetadataCategoryInput:
      type: object
      properties:
        name:
          type: string
      required: [name]
    MetadataCategoryUpdate:
      type: object
      properties:
        name:
          type: string
    MetadataValue:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        value:
          type: string
        mediaCount:
          type: integer
        createdAt:
          type: string
      required: [id, categoryId, value, mediaCount, createdAt]
    MetadataValueUpdate:
      type: object
      properties:
        value:
          type: string
      required: [value]
    MergeValuesInput:
      type: object
      properties:
        sourceIds:
          type: array
          items:
            type: integer
        targetValue:
          type: string
      required: [sourceIds, targetValue]
...
    PlaybackEntry:
      type: object
      properties:
        id:
          type: integer
        mediaId:
          type: integer
        position:
          type: number
        duration:
          type: ["number", "null"]
        completed:
          type: boolean
        playedAt:
          type: string
        media:
          $ref: "#/components/schemas/MediaFile"
      required: [id, mediaId, position, completed, playedAt]
    PlaybackProgressInput:
      type: object
      properties:
        mediaId:
          type: integer
        position:
          type: number
        duration:
          type: ["number", "null"]
        completed:
          type: boolean
      required: [mediaId, position]
...