# lib/api-zod/src/generated/types/metadataCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MediaVault API
 * OpenAPI spec version: 0.1.0
 */

export interface MetadataCategory {
  id: number;
  name: string;
  slug: string;
  isDefault: boolean;
  valueCount: number;
  createdAt: string;
}